+++ packages/opentelemetry-instrumentation-cohere/tests/test_rerank.py
@@ -24,7 +24,7 @@ def test_cohere_rerank(exporter):
         + " in North Dakota in the year 2010. The capital and seat of government is Bismarck.",
     ]
 
-    results = co.rerank(
+    response = co.rerank(
         query=query, documents=documents, top_n=3, model="rerank-multilingual-v2.0"
     )
 
@@ -33,16 +33,22 @@ def test_cohere_rerank(exporter):
     assert cohere_span.name == "cohere.rerank"
     assert cohere_span.attributes.get("gen_ai.system") == "Cohere"
     assert cohere_span.attributes.get("llm.request.type") == "rerank"
-    assert cohere_span.attributes.get("gen_ai.request.model") == "rerank-multilingual-v2.0"
+    assert (
+        cohere_span.attributes.get("gen_ai.request.model") == "rerank-multilingual-v2.0"
+    )
     assert cohere_span.attributes.get(f"gen_ai.prompt.{len(documents)}.role") == "user"
-    assert cohere_span.attributes.get(f"gen_ai.prompt.{len(documents)}.content") == query
+    assert (
+        cohere_span.attributes.get(f"gen_ai.prompt.{len(documents)}.content") == query
+    )
 
     for i, doc in enumerate(documents):
         assert cohere_span.attributes.get(f"gen_ai.prompt.{i}.role") == "system"
         assert cohere_span.attributes.get(f"gen_ai.prompt.{i}.content") == doc
 
-    for idx, result in enumerate(results):
-        assert cohere_span.attributes.get(f"gen_ai.completion.{idx}.role") == "assistant"
+    for idx, result in enumerate(response.results):
+        assert (
+            cohere_span.attributes.get(f"gen_ai.completion.{idx}.role") == "assistant"
+        )
         assert (
             cohere_span.attributes.get(f"gen_ai.completion.{idx}.content")
             == f"Doc {result.index}, Score: {result.relevance_score}\n{result.document['text']}"